backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-restful 
flask-cors
python-dotenv==1.1.0
line-bot-sdk==3.16.3
gunicorn==23.0.0
openai==1.75.0
python-dateutil==2.9.0.post0
stats==0.1.2a0
psycopg2-binary==2.9.6
alembic==1.11.0

sqlalchemy==2.0.20
pydantic==2.5.2
uvicorn==0.23.2
fastapi==0.103.1
greenlet==4.0.0